requirements.txt
flask==2.3.3
pandas==2.0.3
docxtpl==0.16.0
python-docx==1.1.2
openpyxl==3.1.2
gunicorn==21.2.0